requirements.txt
# Core dependencies
pathway==0.5.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.0
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2

# PDF Processing
pymupdf==1.23.7
pdfplumber==0.10.2

# CSV Processing
pandas==2.1.1

# Vector Store
faiss-cpu==1.7.4
chromadb==0.4.18

# Utilities
python-multipart==0.0.6
loguru==0.7.2